Input component reorg + minor code cleanup
- Reorganized input components which have a formik version and base version, so that they're grouped in one file
- Minor code formatting fixes
