+++ packages/shared/src/translations/locales/en/common.json
@@ -78,6 +78,7 @@
     "Calendar": "Calendar",
     "Cancel": "Cancel",
     "Cancelled": "Cancelled",
+    "CanResendIn": "Can resend in {{time}}",
     "Category": "Category",
     "Change": "Change",
     "ChangePassword": "Change Password",
@@ -104,6 +105,7 @@
     "CommentsAsc": "Comments ↑",
     "CommentsDesc": "Comments ↓",
     "Community": "Community",
+    "CompleteVerificationByEnteringCode": "You'll receive a verification code by text. Enter it in the text box below to complete verification.",
     "CompleteVerificationInEmail": "Open our email to complete verification.",
     "Complete": "Complete",
     "Completed": "Completed",
@@ -242,13 +244,14 @@
     "EditLabel": "Edit label",
     "Email_one": "Email",
     "Email_other": "Emails",
+    "EmailDelete": "Delete email",
+    "EmailDeleteConfirm": "Are you sure you want to delete email \"{{emailAddress}}\"?",
+    "EmailListHelp": "This list contains all of your connected email addresses.\n\nYou may add or remove as many email addresses as you wish, but you must keep at least one *verified* authentication method (either a phone number, email address, or crypto wallet).",
     "EmailNotification_one": "Email Notification",
     "EmailNotification_other": "Email Notifications",
     "EmailNotificationToggleDescription": "Toggle all email notifications on or off.",
-    "EmailPlaceholder": "Enter email...",
-    "EmailDeleteConfirm": "Are you sure you want to delete email \"{{emailAddress}}\"?",
-    "EmailListHelp": "This list contains all of your connected email addresses.\n\nYou may add or remove as many email addresses as you wish, but you must keep at least one *verified* authentication method (either a wallet or email address).",
     "EmailLogin": "Email Login",
+    "EmailPlaceholder": "Enter email...",
     "EmailVerified": "Email verified!",
     "EmojisSuggested": "Recently Used",
     "EmojisSmileysPeople": "Smileys & People",
@@ -464,6 +467,7 @@
     "New": "New",
     "NewChat": "New chat",
     "NewEmailAddress": "New email address",
+    "NewPhoneNumber": "New phone number",
     "Next": "Next",
     "No": "No",
     "None": "None",
@@ -551,6 +555,13 @@
     "Performance": "Performance",
     "Persona": "Persona",
     "PersonaPlaceholderBot": "How would you describe the personality of this bot?",
+    "Phone_one": "Phone",
+    "Phone_other": "Phones",
+    "PhoneDelete": "Delete phone",
+    "PhoneDeleteConfirm": "Are you sure you want to delete phone number \"{{phoneNumber}}\"?",
+    "PhoneListHelp": "This list contains all of your connected phone numbers.\n\nYou may add or remove numbers, but you must keep at least one *verified* authentication method (either a phone number, email address, or crypto wallet).",
+    "PhoneNumberPlaceholder": "Enter phone number...",
+    "PhoneVerified": "Phone verified!",
     "PleaseBeNice": "Please be nice to each other...",
     "PreviewTo": "To preview",
     "Premium": "Premium",
@@ -647,6 +658,8 @@
     "ReportSubmitted": "Report submitted.",
     "RequestSentCheckEmail": "Request sent. Check your email.",
     "Researching": "Researching",
+    "ResendVerificationEmail": "Resend verification email",
+    "ResendVerificationText": "Resend verification text",
     "Reset": "Reset",
     "ResetPassword": "Reset password",
     "Resource_one": "Resource",
@@ -966,8 +979,13 @@
     "UserTogglePressEditable": "User: {{user}}. Press to remove",
     "Verbosity": "Verbosity",
     "VerbosityPlaceholder": "How long should this bot's responses be?",
+    "VerificationCode": "Verification code",
+    "VerificationCodePlaceholder": "12345678",
+    "Verified": "Verified",
+    "VerifiedNot": "Not verified",
     "VerifiedEmailsCreated": "Emails verified",
     "VerifiedWalletsCreated": "Wallets verified",
+    "VerifyWallet": "Verify wallet",
     "VersionsAsc": "Versions ↑",
     "VersionsDesc": "Versions ↓",
     "VersionAndRootComplete": "Version and root complete?",
@@ -992,8 +1010,9 @@
     "VotesHelp": "Votes are a way to show support for an object, which affect the ranking of an object in default searches.",
     "Wallet_one": "Wallet",
     "Wallet_other": "Wallets",
+    "WalletDelete": "Delete wallet",
     "WalletDeleteConfirm": "Are you sure you want to delete wallet \"{{walletName}}\"?",
-    "WalletListHelp": "This list contains all of your connected wallets. If a custom name has not been set, the wallet's reward address will be displayed.\n\nYou may add or remove as many wallets as you wish, but you must keep at least one *verified* authentication method (either a wallet or email address).",
+    "WalletListHelp": "This list contains all of your connected wallets. If a custom name has not been set, the wallet's reward address will be displayed.\n\nYou may add or remove as many wallets as you wish, but you must keep at least one *verified* authentication method (either a phone number, email address, or crypto wallet).",
     "WalletVerified": "Wallet verified!",
     "WebContent": "Web content",
     "Week": "Week",
