+++ packages/shared/src/translations/locales/en/error.json
@@ -16,6 +16,8 @@
     "CannotOpenLink": "Failed to open link.",
     "CannotVerifyEmailCode": "Cannot verify code.",
     "CannotVerifyEmailCodeDetails": "Email verification code is invalid.",
+    "CannotVerifyPhoneCode": "Cannot verify code.",
+    "CannotVerifyPhoneCodeDetails": "Phone verification code is invalid.",
     "CannotVerifyWallet": "Cannot verify wallet.",
     "CannotVerifyWalletDetails": "Wallet verification failed. Please try again.",
     "ChatRoomJoinFailed": "Failed to join chat room.",
@@ -29,6 +31,7 @@
     "CreateNotSupported": "Cannot create this type of object",
     "DeleteCommentFailed": "Failed to delete comment.",
     "DeleteNotSupported": "Cannot delete this type of object",
+    "EmailAlreadyVerified": "Email already verified.",
     "EmailInUse": "Email already in use.",
     "EmailInUseDetails": "An account with this email already exists.",
     "EmailNotFound": "Email not found.",
@@ -80,7 +83,7 @@
     "MaxReportsReachedDetails": "You can only have one open report on an object at a time.",
     "MustHaveFocusMode": "Must have at least one focus mode.",
     "MustLeaveVerificationMethod": "Cannot delete all verified authentication methods.",
-    "MustLeaveVerificationMethodDetails": "You must have at least one verified authentication method (email or wallet). Please verify another method before deleting this one.",
+    "MustLeaveVerificationMethodDetails": "You must have at least one verified authentication method (phone number, email address, or crypto wallet). Please verify another method before deleting this one.",
     "MustResetPassword": "Must reset password.",
     "MustResetPasswordDetails": "Before signing in, please follow the link sent to your email to change your password.",
     "MustUseApiToken": "Call must be made from an API token.",
@@ -96,6 +99,7 @@
     "NoFilesSelected": "No files selected.",
     "NoFocusModes": "No focus modes",
     "NoInternet": "No internet connection.",
+    "NoPhoneNumbers": "No phone numbers found.",
     "NoPushDevices": "No push devices found.",
     "NoResults": "No results",
     "NoTextSelected": "No text selected.",
@@ -122,8 +126,13 @@
     "PageNotFoundDetails": "Looks like you've followed a broken link or entered a URL that doesn't exist on this site.",
     "PageRestricted": "Page restricted. Please log in.",
     "PasswordsDontMatch": "Passwords don't match.",
+    "PhoneAlreadyVerified": "Phone already verified.",
+    "PhoneCodeSentRecently": "Code sent too recently.",
     "PhoneInUse": "Phone number taken.",
     "PhoneInUseDetails": "Phone number has already been verified by another account. If you believe this is a mistake, please contact us.",
+    "PhoneNotFound": "Phone number not found.",
+    "PhoneNotFoundDetails": "No account with this phone number exists.",
+    "PhoneNotYours": "Phone number not yours.",
     "PushDeviceNotFound": "Push device not found.",
     "PushDeviceNotFoundDetails": "Could not find push device, or the device does not belong to you.",
     "RateLimitExceeded": "Rate limit exceeded.",
