+++ packages/shared/src/validation/utils/versions/index.ts
@@ -1,5 +1,6 @@
 export * from "./calculateVersionsFromString";
-export * from './getMinVersion';
+export * from "./getMinVersion";
 export * from "./meetsMinVersion";
 export * from "./minVersionTest";
+
