+++ packages/ui/src/components/inputs/RichInputLexical/builder.ts
@@ -45,7 +45,7 @@ function findOutermostMatch(textContent, textTransformersIndex) {
     for (const match of openTagsMatch) {
         // Open tags reg exp might capture leading space so removing it
         // before using match to find transformer
-        const tag = match.replace(/^\s/, '');
+        const tag = match.replace(/^\s/, "");
         const fullMatchRegExp = textTransformersIndex.fullMatchRegExpByTag[tag];
         if (fullMatchRegExp == null) {
             continue;
@@ -60,7 +60,7 @@ function findOutermostMatch(textContent, textTransformersIndex) {
             // For non-intraword transformers checking if it's within a word
             // or surrounded with space/punctuation/newline
             const {
-                index = 0
+                index = 0,
             } = fullMatch;
             const beforeChar = textContent[index - 1];
             const afterChar = textContent[index + fullMatch[0].length];
@@ -117,7 +117,7 @@ const importTextFormatTransformers = (textNode, textFormatTransformersIndex, tex
     }
 
     // Recursively run over inner text if it's not inline code
-    if (!currentNode.hasFormat('code')) {
+    if (!currentNode.hasFormat("code")) {
         importTextFormatTransformers(currentNode, textFormatTransformersIndex, textMatchTransformers);
     }
 
@@ -128,7 +128,7 @@ const importTextFormatTransformers = (textNode, textFormatTransformersIndex, tex
     if (remainderNode) {
         importTextFormatTransformers(remainderNode, textFormatTransformersIndex, textMatchTransformers);
     }
-}
+};
 
 const importTextMatchTransformers = (textNode_, textMatchTransformers) => {
     let textNode = textNode_;
@@ -154,7 +154,7 @@ const importTextMatchTransformers = (textNode_, textMatchTransformers) => {
         }
         break;
     }
-}
+};
 
 const importBlocks = (lineText, rootNode, elementTransformers, textFormatTransformersIndex, textMatchTransformers) => {
     const lineTextTrimmed = lineText.trim();
@@ -164,7 +164,7 @@ const importBlocks = (lineText, rootNode, elementTransformers, textFormatTransfo
     rootNode.append(elementNode);
     for (const {
         regExp,
-        replace
+        replace,
     } of elementTransformers) {
         const match = lineText.match(regExp);
         if (match) {
@@ -196,7 +196,7 @@ const importBlocks = (lineText, rootNode, elementTransformers, textFormatTransfo
             }
         }
     }
-}
+};
 
 /**
  * Imports a code block from markdown content into a Lexical document.
@@ -225,7 +225,7 @@ const importCodeBlock = (lines: string[], startLineIndex: number, rootNode: Elem
             const closeMatch = lines[endLineIndex].match(CODE_BLOCK_REG_EXP);
             if (closeMatch) {
                 const codeBlockNode = $createCodeBlockNode(openMatch[1]);
-                const textNode = $createTextNode(lines.slice(startLineIndex + 1, endLineIndex).join('\n'));
+                const textNode = $createTextNode(lines.slice(startLineIndex + 1, endLineIndex).join("\n"));
                 codeBlockNode.append(textNode);
                 rootNode.append(codeBlockNode);
                 return [codeBlockNode, endLineIndex];
@@ -233,7 +233,7 @@ const importCodeBlock = (lines: string[], startLineIndex: number, rootNode: Elem
         }
     }
     return [null, startLineIndex];
-}
+};
 
 /**
  * Organizes items from a list into an index (a Record) based on a key generated by a callback function.
@@ -262,14 +262,14 @@ const transformersByType = (transformers: Transformer[]) => {
     const byType = indexBy(transformers, t => t.type);
     return {
         element: byType.element || [],
-        textFormat: byType['text-format'] || [],
-        textMatch: byType['text-match'] || []
+        textFormat: byType["text-format"] || [],
+        textMatch: byType["text-match"] || [],
     } as {
         element: ElementTransformer[];
         textFormat: TextFormatTransformer[];
         textMatch: TextMatchTransformer[];
     };
-}
+};
 
 const isEmptyParagraph = (node) => {
     if (!$isParagraphNode(node)) {
@@ -277,7 +277,7 @@ const isEmptyParagraph = (node) => {
     }
     const firstChild = node.getFirstChild();
     return firstChild == null || node.getChildrenSize() === 1 && $isTextNode(firstChild) && MARKDOWN_EMPTY_LINE_REG_EXP.test(firstChild.getTextContent());
-}
+};
 
 /**
  * Creates an index of text transformers based on the provided array of text transformers.
@@ -305,13 +305,13 @@ const createTextFormatTransformersIndex = (textTransformers: TextFormatTransform
     const transformersByTag = {};
     const fullMatchRegExpByTag = {};
     const openTagsRegExp: string[] = [];
-    const escapeRegExp = `(?<![\\\\])`;
+    const escapeRegExp = "(?<![\\\\])";
     const { deviceOS } = getDeviceInfo();
     const isApple = deviceOS === DeviceOS.IOS || deviceOS === DeviceOS.MacOS;
     for (const transformer of textTransformers) {
         const { tag } = transformer;
         transformersByTag[tag] = transformer;
-        const tagRegExp = tag.replace(/(\*|\^|\+)/g, '\\$1');
+        const tagRegExp = tag.replace(/(\*|\^|\+)/g, "\\$1");
         openTagsRegExp.push(tagRegExp);
         if (isApple) {
             fullMatchRegExpByTag[tag] = new RegExp(`(${tagRegExp})(?![${tagRegExp}\\s])(.*?[^${tagRegExp}\\s])${tagRegExp}(?!${tagRegExp})`);
@@ -323,10 +323,10 @@ const createTextFormatTransformersIndex = (textTransformers: TextFormatTransform
         // Reg exp to find open tag + content + close tag
         fullMatchRegExpByTag,
         // Reg exp to find opening tags
-        openTagsRegExp: new RegExp((isApple ? '' : `${escapeRegExp}`) + '(' + openTagsRegExp.join('|') + ')', 'g'),
-        transformersByTag
+        openTagsRegExp: new RegExp((isApple ? "" : `${escapeRegExp}`) + "(" + openTagsRegExp.join("|") + ")", "g"),
+        transformersByTag,
     };
-}
+};
 
 /**
  * Creates a function that imports markdown content into a Lexical document.
@@ -353,7 +353,7 @@ const createMarkdownImport = (transformers: Transformer[]): ((markdownString: st
     const byType = transformersByType(transformers);
     const textFormatTransformersIndex = createTextFormatTransformersIndex(byType.textFormat);
     return (markdownString, node) => {
-        const lines = markdownString.split('\n');
+        const lines = markdownString.split("\n");
         const linesLength = lines.length;
         const root = node || $getRoot();
         root.clear();
@@ -373,7 +373,7 @@ const createMarkdownImport = (transformers: Transformer[]): ((markdownString: st
             root.selectEnd();
         }
     };
-}
+};
 
 /**
  * Get next or previous text sibling a text node, including cases 
@@ -407,11 +407,11 @@ const getTextSibling = (node, backward) => {
         }
     }
     return null;
-}
+};
 
 const hasFormat = (node, format) => {
     return $isTextNode(node) && node.hasFormat(format);
-}
+};
 
 const exportTextFormat = (node: LexicalNode, textContent: string, textFormatTransformers: TextFormatTransformer[]) => {
     // This function handles the case of a string looking like this: "   foo   "
@@ -443,7 +443,7 @@ const exportTextFormat = (node: LexicalNode, textContent: string, textFormatTran
 
     // Replace trimmed version of textContent ensuring surrounding whitespace is not modified
     return textContent.replace(frozenString, output);
-}
+};
 
 const exportChildren = (node: ElementNode, textFormatTransformers: TextFormatTransformer[], textMatchTransformers: TextMatchTransformer[]) => {
     const output: string[] = [];
@@ -467,7 +467,7 @@ const exportChildren = (node: ElementNode, textFormatTransformers: TextFormatTra
         }
     }
     return output.join("");
-}
+};
 
 const exportTopLevelElements = (
     node: LexicalNode,
@@ -488,7 +488,7 @@ const exportTopLevelElements = (
     } else {
         return null;
     }
-}
+};
 
 /**
  * Creates a function that exports Lexical document content or content within a specified node to markdown format.
@@ -527,9 +527,9 @@ const createMarkdownExport = (transformers: Transformer[]): ((node?: ElementNode
                 output.push(result);
             }
         }
-        return output.join('\n\n');
+        return output.join("\n\n");
     };
-}
+};
 
 /**
  * Converts a markdown string to Lexical nodes using the specified transformers.
@@ -547,9 +547,9 @@ const createMarkdownExport = (transformers: Transformer[]): ((node?: ElementNode
  */
 export const $convertFromMarkdownString = (markdown: string, transformers: Transformer[], node?: ElementNode) => {
     const importMarkdown = createMarkdownImport(transformers);
-    console.log('got importMarkdown');
+    console.log("got importMarkdown");
     return importMarkdown(markdown, node); //TODO might not need a return
-}
+};
 
 /**
  * Converts Lexical nodes within a specified parent node or the entire document into a markdown string
@@ -570,4 +570,4 @@ export const $convertFromMarkdownString = (markdown: string, transformers: Trans
 export const $convertToMarkdownString = (transformers: Transformer[], node?: ElementNode) => {
     const exportMarkdown = createMarkdownExport(transformers);
     return exportMarkdown(node);
-}
+};
