+++ packages/ui/src/tools/api/partial/user.ts
@@ -94,6 +94,7 @@ export const profile: GqlPartial<User> = {
         theme: true,
         emails: async () => rel((await import("./email")).email, "full"),
         focusModes: async () => rel((await import("./focusMode")).focusMode, "full"),
+        phones: async () => rel((await import("./phone")).phone, "full"),
         pushDevices: async () => rel((await import("./pushDevice")).pushDevice, "full"),
         wallets: async () => rel((await import("./wallet")).wallet, "common"),
         notifications: async () => rel((await import("./notification")).notification, "full"),
