+++ packages/ui/src/views/objects/memberInvite/MemberInvitesUpsert/MemberInvitesUpsert.tsx
@@ -3,7 +3,7 @@ import { Box, Checkbox, FormControlLabel, Typography, useTheme } from "@mui/mate
 import { fetchLazyWrapper } from "api";
 import { BottomActionsButtons } from "components/buttons/BottomActionsButtons/BottomActionsButtons";
 import { MaybeLargeDialog } from "components/dialogs/LargeDialog/LargeDialog";
-import { RichInputBase } from "components/inputs/RichInputBase/RichInputBase";
+import { RichInputBase } from "components/inputs/RichInput/RichInput";
 import { ObjectList } from "components/lists/ObjectList/ObjectList";
 import { TopBar } from "components/navigation/TopBar/TopBar";
 import { Field, Formik } from "formik";
