+++ packages/ui/src/views/objects/resource/ResourceUpsert/ResourceUpsert.tsx
@@ -4,8 +4,7 @@ import { fetchLazyWrapper } from "api";
 import { BottomActionsButtons } from "components/buttons/BottomActionsButtons/BottomActionsButtons";
 import { MaybeLargeDialog } from "components/dialogs/LargeDialog/LargeDialog";
 import { LinkInput } from "components/inputs/LinkInput/LinkInput";
-import { Selector } from "components/inputs/Selector/Selector";
-import { SelectorBase } from "components/inputs/SelectorBase/SelectorBase";
+import { Selector, SelectorBase } from "components/inputs/Selector/Selector";
 import { TranslatedTextInput } from "components/inputs/TranslatedTextInput/TranslatedTextInput";
 import { TopBar } from "components/navigation/TopBar/TopBar";
 import { SessionContext } from "contexts/SessionContext";
